feat: make helm repo index point at GitHub release
Replace the call to `helm repo index` with a ruby script that works from
the concourse artifacts, so that we do not need to upload the chart
bundles into git; instead, we build the GitHub release URL and use that
instead.  Also switch back to having the time stamp in the artifact
name, so that we can still identify the different build artifacts (but
only use the newest one in the chart index).