fix: Attempt to publish repo via github actions
Having issues with the github-release concourse resource (it's not
triggering builds when assets are uploaded); trying with github actions
instead.

Ends up we also need to have a release body text to force the change to
be recognized as an edit.  Put in the file name as we need to regenerate
the index whenever the file name changes.