Support Extensions as reconcilers
We now support reconcilers that are satisfying the eirinix.Reconciler interface.
Reconcilers can act as a normal extension, but with a broader set of options.

To note, reconcilers are loaded just like eirinix.Extensions, so the execution flow of
EiriniX doesn't change, instead we can now have both Extensions and Reconcilers running
in the same process.