Allow to start Watchers and Reconcilers in eirinix.Start()
In this way there isn't anymore difference for the end user on how to start
the eirinix Manager. Now Reconcilers, Watchers and Extensions can be added
with `eirinix.AddExtension()`.

This also means that in future we could deprecate the `Watch()` interface and
the single method used for the specific types.

To note, `AddExtension` need pointers to structures.